Move Allegiance to its own class
