+++ Tests/Hypothesis/testAllegiance.py
@@ -7,7 +7,7 @@ from datetime import timedelta
 from hypothesis import given, assume, example, HealthCheck, settings
 from hypothesis.strategies import text, from_regex, composite, integers, floats, lists, sampled_from, booleans
 
-from PyRoute.Galaxy import Allegiance
+from PyRoute.AreaItems.Allegiance import Allegiance
 
 
 @composite
